ref(bottle-classifier): Extract exact-cask scope policy
Move generic exact-cask scope detection out of the main review policy so the review layer stays focused on decision orchestration. This keeps deterministic cask-scope rules in a small module without changing behavior.
